+++ sidebars.js
@@ -43,91 +43,90 @@ const sidebars = {
             "developer-docs/build/project-setup/cycles-wallet",
           ],
         },
-        "developer-docs/build/using-an-agent",
         {
           type: "category",
-          label: "Languages",
+          label: "Agents",
           link: {
             type: "doc",
-            id: "developer-docs/build/languages/work-with-languages",
+            id: "developer-docs/build/agents/index",
           },
           items: [
-            // 'developer-docs/build/explore-templates',
             {
               type: "category",
-              label: "Rust",
+              label: "JavaScript Agent by DFINITY",
               link: {
                 type: "doc",
-                id: "developer-docs/build/languages/rust/rust-intro",
+                id: "developer-docs/build/agents/javascript/javascript-intro",
               },
               items: [
-                "developer-docs/build/languages/rust/rust-quickstart",
-                "developer-docs/build/languages/rust/rust-counter",
-                "developer-docs/build/languages/rust/multiply-dependency",
-                "developer-docs/build/languages/rust/rust-profile",
-                "developer-docs/build/languages/rust/rust-optimize",
+                "developer-docs/build/agents/javascript/javascript-intro",
+                "developer-docs/build/agents/javascript/nodejs",
+              ],
+            },
+            "developer-docs/build/agents/ic-agent-dfinity",
+          ],
+        },
+        {
+          type: "category",
+          label: "CDKs",
+          link: {
+            type: "doc",
+            id: "developer-docs/build/cdks/index",
+          },
+          items: [
+            {
+              type: "category",
+              label: "cdk-rs by DFINITY (Rust)",
+              link: {
+                type: "doc",
+                id: "developer-docs/build/cdks/cdk-rs-dfinity/index",
+              },
+              items: [
+                "developer-docs/build/cdks/cdk-rs-dfinity/rust-quickstart",
+                "developer-docs/build/cdks/cdk-rs-dfinity/rust-profile",
+                "developer-docs/build/cdks/cdk-rs-dfinity/rust-counter",
+                "developer-docs/build/cdks/cdk-rs-dfinity/rust-optimize",
+                "developer-docs/build/cdks/cdk-rs-dfinity/multiply-dependency",
               ],
             },
             {
               type: "category",
-              label: "Motoko",
-              link: {
-                // type: 'doc', id: 'developer-docs/build/languages/motoko/index',
-                type: "doc",
-                id: "developer-docs/build/languages/motoko/motoko",
-              },
-              items: [
-                "developer-docs/build/languages/motoko/about-this-guide",
-                "developer-docs/build/languages/motoko/motoko-introduction",
-                "developer-docs/build/languages/motoko/basic-concepts",
-                "developer-docs/build/languages/motoko/base-intro",
-                "developer-docs/build/languages/motoko/mutable-state",
-                "developer-docs/build/languages/motoko/local-objects-classes",
-                "developer-docs/build/languages/motoko/actors-async",
-                "developer-docs/build/languages/motoko/errors",
-                "developer-docs/build/languages/motoko/pattern-matching",
-                "developer-docs/build/languages/motoko/sharing",
-                "developer-docs/build/languages/motoko/message-inspection",
-                "developer-docs/build/languages/motoko/modules-and-imports",
-                "developer-docs/build/languages/motoko/control-flow",
-                "developer-docs/build/languages/motoko/structural-equality",
-                "developer-docs/build/languages/motoko/actor-classes",
-                "developer-docs/build/languages/motoko/caller-id",
-                "developer-docs/build/languages/motoko/cycles",
-                "developer-docs/build/languages/motoko/upgrades",
-                "developer-docs/build/languages/motoko/compatibility",
-                "developer-docs/build/languages/motoko/stablememory",
-                "developer-docs/build/languages/motoko/heartbeats",
-                "developer-docs/build/languages/motoko/language-manual",
-                "developer-docs/build/languages/motoko/compiler-ref",
-                "developer-docs/build/languages/motoko/motoko-grammar",
-                "developer-docs/build/languages/motoko/overview",
-                "developer-docs/build/languages/motoko/style",
-              ],
-            },
-            {
-              type: "category",
-              label: "Candid",
+              label: "Motoko by DFINITY",
               link: {
                 type: "doc",
-                id: "developer-docs/build/languages/candid/candid-intro",
+                id: "developer-docs/build/cdks/motoko-dfinity/motoko",
               },
               items: [
-                "developer-docs/build/languages/candid/candid-concepts",
-                "developer-docs/build/languages/candid/candid-howto",
+                "developer-docs/build/cdks/motoko-dfinity/about-this-guide",
+                "developer-docs/build/cdks/motoko-dfinity/motoko-introduction",
+                "developer-docs/build/cdks/motoko-dfinity/basic-concepts",
+                "developer-docs/build/cdks/motoko-dfinity/base-intro",
+                "developer-docs/build/cdks/motoko-dfinity/mutable-state",
+                "developer-docs/build/cdks/motoko-dfinity/local-objects-classes",
+                "developer-docs/build/cdks/motoko-dfinity/actors-async",
+                "developer-docs/build/cdks/motoko-dfinity/errors",
+                "developer-docs/build/cdks/motoko-dfinity/pattern-matching",
+                "developer-docs/build/cdks/motoko-dfinity/sharing",
+                "developer-docs/build/cdks/motoko-dfinity/message-inspection",
+                "developer-docs/build/cdks/motoko-dfinity/modules-and-imports",
+                "developer-docs/build/cdks/motoko-dfinity/control-flow",
+                "developer-docs/build/cdks/motoko-dfinity/structural-equality",
+                "developer-docs/build/cdks/motoko-dfinity/actor-classes",
+                "developer-docs/build/cdks/motoko-dfinity/caller-id",
+                "developer-docs/build/cdks/motoko-dfinity/cycles",
+                "developer-docs/build/cdks/motoko-dfinity/upgrades",
+                "developer-docs/build/cdks/motoko-dfinity/compatibility",
+                "developer-docs/build/cdks/motoko-dfinity/stablememory",
+                "developer-docs/build/cdks/motoko-dfinity/heartbeats",
+                "developer-docs/build/cdks/motoko-dfinity/language-manual",
+                "developer-docs/build/cdks/motoko-dfinity/compiler-ref",
+                "developer-docs/build/cdks/motoko-dfinity/motoko-grammar",
+                "developer-docs/build/cdks/motoko-dfinity/overview",
+                "developer-docs/build/cdks/motoko-dfinity/style",
               ],
             },
-            {
-              type: "category",
-              label: "Other Languages",
-              link: {
-                type: "doc",
-                id: "developer-docs/build/languages/other-languages/other-languages-intro",
-              },
-              items: [
-                "developer-docs/build/languages/other-languages/clang-supported-languages",
-              ],
-            },
+            "developer-docs/build/cdks/clang-supported-languages",
+            "developer-docs/build/cdks/azle-demergent-labs",
           ],
         },
         {
@@ -165,6 +164,18 @@ const sidebars = {
             "developer-docs/build/frontend/custom-frontend",
           ],
         },
+        {
+          type: "category",
+          label: "Candid",
+          link: {
+            type: "doc",
+            id: "developer-docs/build/candid/candid-intro",
+          },
+          items: [
+            "developer-docs/build/candid/candid-concepts",
+            "developer-docs/build/candid/candid-howto",
+          ],
+        },
         "developer-docs/build/troubleshooting",
       ],
     },
@@ -173,12 +184,12 @@ const sidebars = {
       label: "Deployment & Scaling",
       link: {
         type: "doc",
-        id: "developer-docs/deploy/computation-and-storage-costs",
+        id: "developer-docs/deploy/deploy",
       },
       items: [
-        // 'developer-docs/deploy/deploy',
-        // 'developer-docs/deploy/scale',
+        "developer-docs/deploy/deploy",
         "developer-docs/deploy/larger-wasm",
+        "developer-docs/deploy/staging-environment",
         "developer-docs/deploy/custom-domain",
         "developer-docs/deploy/computation-and-storage-costs",
         // 'developer-docs/deploy/advanced-deployment'
@@ -186,70 +197,80 @@ const sidebars = {
     },
     {
       type: "category",
-      label: "Integrate Functionality",
+      label: "Integrations",
       link: {
-        // type: 'doc', id: 'developer-docs/functionality/index'
-        type: "doc",
-        id: "developer-docs/functionality/bitcoin/index",
+        type: 'doc', id: 'developer-docs/integrations/bitcoin/index'
       },
       items: [
-        // 'developer-docs/functionality/index',
+        // 'developer-docs/integrations/index',
+        // {
+        //   type: 'category',
+        //   label: 'Service Nervous System (SNS)',
+        //   link: {
+        //     type: "doc",
+        //     id: "developer-docs/integrations/sns/index",
+        //   },
+        //   items: [
+        //     "developer-docs/integrations/sns/sns-intro",
+        //     "developer-docs/integrations/sns/deployment/launch-intro",
+        //     "developer-docs/integrations/sns/deployment/frontend-integration",
+        //     "developer-docs/integrations/sns/deployment/predeployment",
+        //     "developer-docs/integrations/sns/deployment/local-testing",
+        //     "developer-docs/integrations/sns/deployment/deployment",
+        //     "developer-docs/integrations/sns/managing-sns/nervous-system-parameters",
+        //     "developer-docs/integrations/sns/managing-sns/proposal-guide",
+        //     "developer-docs/integrations/sns/managing-sns/cycles-usage",
+        //     "developer-docs/integrations/sns/managing-sns/upgradeSNS",
+        //   ],
+        // },
         {
-          type: 'category',
-          label: 'Bitcoin Integration',
+          type: "category",
+          label: "Bitcoin Integration",
           link: {
             type: "doc",
-            id: "developer-docs/functionality/bitcoin/index"
+            id: "developer-docs/integrations/bitcoin/index",
           },
           items: [
-            'developer-docs/functionality/bitcoin/bitcoin-how-it-works',
-            'developer-docs/functionality/bitcoin/local-development',
-          ]
+            "developer-docs/integrations/bitcoin/bitcoin-how-it-works",
+            "developer-docs/integrations/bitcoin/local-development",
+          ],
         },
         {
-          type: 'category',
-          label: 'Threshold ECDSA',
+          type: "category",
+          label: "Threshold ECDSA",
           link: {
             type: "doc",
-            id: "developer-docs/functionality/t-ecdsa/index"
+            id: "developer-docs/integrations/t-ecdsa/index",
           },
-          items: [
-            'developer-docs/functionality/t-ecdsa/t-ecdsa-how-it-works'
-          ]
+          items: ["developer-docs/integrations/t-ecdsa/t-ecdsa-how-it-works"],
         },
-        "developer-docs/functionality/internet-identity/integrate-identity",
+        "developer-docs/integrations/internet-identity/integrate-identity",
         {
           type: "category",
           label: "ICP Ledger",
           link: {
             type: "doc",
-            id: "developer-docs/functionality/ledger/index",
+            id: "developer-docs/integrations/ledger/index",
           },
           items: [
-            "developer-docs/functionality/ledger/interact-with-ledger",
-            "developer-docs/functionality/ledger/ledger-local-setup",
-            "developer-docs/functionality/ledger/deploy-new-token",
+            "developer-docs/integrations/ledger/interact-with-ledger",
+            "developer-docs/integrations/ledger/ledger-local-setup",
+            "developer-docs/integrations/ledger/deploy-new-token",
           ],
         },
-        // {
-        //   type: 'category',
-        //   label: 'Using the SNS',
-        //   items: [
-        //     'developer-docs/functionality/sns',
-        //   ]
-        // },
         {
           type: "category",
           label: "Rosetta",
           link: {
             type: "doc",
-            id: "developer-docs/functionality/rosetta/index",
+            id: "developer-docs/integrations/rosetta/index",
           },
           items: [
-            "developer-docs/functionality/rosetta/transfers",
-            "developer-docs/functionality/rosetta/neuron-lifecycle",
-            "developer-docs/functionality/rosetta/staking-support",
-            "developer-docs/functionality/rosetta/staking-tutorial",
+            "developer-docs/integrations/rosetta/transfers",
+            "developer-docs/integrations/rosetta/neuron-lifecycle",
+            "developer-docs/integrations/rosetta/staking-support",
+            "developer-docs/integrations/rosetta/staking-tutorial",
+            "developer-docs/integrations/rosetta/hotkeys",
           ],
         },
       ],
@@ -309,7 +330,7 @@ const sidebars = {
     {
       type: "doc",
       label: "dfx.json schema",
-      id: "references/dfx-json-reference"
+      id: "references/dfx-json-reference",
     },
     {
       type: "category",
@@ -430,10 +451,27 @@ const sidebars = {
     "concepts/trust-in-canisters",
     "concepts/tokens-cycles",
     "concepts/governance",
-    "concepts/bitcoin-integration",
   ],
   tokenomics: [
     "tokenomics/index",
+    // {
+    //   type: "category",
+    //   label: "SNS",
+    //   link:{
+    //     type: "doc",
+    //     id: "tokenomics/sns/intro"
+    //   },
+    //   items: [
+    //     "tokenomics/sns/intro",
+    //     "tokenomics/sns/rewards",
+    //     "tokenomics/sns/community-fund",
+    //     "tokenomics/sns/not-blind/predeployment-considerations",
+    //     "tokenomics/sns/not-blind/tokenomics",
+    //     "tokenomics/sns/not-blind/product",
+    //     "tokenomics/sns/not-blind/business",
+    //     "tokenomics/sns/not-blind/hype"
+    //   ],
+    // },
     {
       type: "category",
       label: "Identity & Authentication",