+++ src/components/BitcoinIntegrationPage/Hero/index.tsx
@@ -2,8 +2,7 @@ import React from "react";
 import { motion } from "framer-motion";
 import transitions from "@site/static/transitions.json";
 import AnimateSpawn from "@site/src/components/Common/AnimateSpawn";
-
-const css = (strings) => strings.join("");
+import { css } from "@site/src/utils/dummy-css";
 
 const Hero = () => {
   return (
@@ -57,17 +56,17 @@ const Hero = () => {
           variants={transitions.item}
           className="tw-heading-3 mb-2 w-5/6 md:tw-heading-2 md:mb-6 md:ml-1/12 md:w-6/12"
         >
-          Bitcoin Integration
+          Code Bitcoin
         </motion.h1>
         <motion.p
           variants={transitions.item}
           className="tw-lead-sm md:tw-lead md:ml-1/12 md:w-7/12"
         >
-          The Internet Computer enables direct integration with the Bitcoin network. 
-          By way of bridge-less communication with the Bitcoin network and a novel
-          threshold ECDSA protocol, canisters on the Internet Computer can now 
-          securely receive, hold, and send bitcoins.
-
+          The Internet Computer is cryptographically integrated with the Bitcoin
+          network. The canister smart contracts it hosts can create bitcoin addresses, and send and receive bitcoin,
+          directly on the Bitcoin ledger. This means "wrapped bitcoin" from insecure centralized
+          "bridge" services, which have been hacked for billions of dollars, are no longer needed
+          in DeFi and Web3. Say hello to <b>1-chain-is-many-chains</b>!
         </motion.p>
         <svg
           viewBox="0 0 917 830"