+++ src/utils/dummy-css.ts
@@ -1,0 +1,2 @@
+// for editor syntax highlight only
+export const css = (strings) => strings.join("");